+++ src/main/java/codesquad/springcafe/articles/repository/ArticleRepository.java
@@ -1,10 +1,7 @@
 package codesquad.springcafe.articles.repository;
 
 import codesquad.springcafe.articles.model.Article;
-import codesquad.springcafe.articles.model.Reply;
-import codesquad.springcafe.articles.model.dto.ArticleCreationRequest;
 import codesquad.springcafe.articles.model.dto.ArticleUpdateDto;
-import codesquad.springcafe.articles.model.dto.ReplyCreationRequest;
 
 import java.util.ArrayList;
 import java.util.Optional;
@@ -21,12 +18,4 @@ public interface ArticleRepository {
     void updateArticle(long articleId, ArticleUpdateDto articleUpdateDto);
 
     void deleteArticle(long articleId);
-
-    Reply createReply(Reply reply);
-
-    Optional<ArrayList<Reply>> getReplies(long articleId);
-
-    Optional<Reply> findReplyById(long replyId);
-
-    boolean deleteReply(long replyId);
 }
