+++ src/main/java/codesquad/springcafe/config/interceptor/LoggedInInterceptor.java
@@ -3,8 +3,6 @@ package codesquad.springcafe.config.interceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 public class LoggedInInterceptor implements HandlerInterceptor {
