+++ src/main/java/codesquad/springcafe/replies/controller/ReplyRestController.java
@@ -1,9 +1,9 @@
 package codesquad.springcafe.replies.controller;
 
-import codesquad.springcafe.articles.model.Reply;
-import codesquad.springcafe.articles.model.dto.ReplyCreationRequest;
-import codesquad.springcafe.articles.model.dto.ReplyViewDto;
-import codesquad.springcafe.articles.service.ArticleService;
+import codesquad.springcafe.replies.model.Reply;
+import codesquad.springcafe.replies.model.dto.ReplyCreationRequest;
+import codesquad.springcafe.replies.model.dto.ReplyViewDto;
+import codesquad.springcafe.replies.service.ReplyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,16 +11,16 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/api/articles/{articleId}/replies")
 public class ReplyRestController {
-    private final ArticleService articleService;
+    private final ReplyService replyService;
 
     @Autowired
-    public ReplyRestController(ArticleService articleService) {
-        this.articleService = articleService;
+    public ReplyRestController(ReplyService replyService) {
+        this.replyService = replyService;
     }
 
     @PostMapping
     public ResponseEntity<ReplyViewDto> createReply(@PathVariable long articleId, @RequestBody ReplyCreationRequest replyCreationRequest) {
-        Reply reply = articleService.createReply(articleId, replyCreationRequest);
+        Reply reply = replyService.createReply(articleId, replyCreationRequest);
 
         ReplyViewDto replyViewDto = new ReplyViewDto(reply, true);
 
@@ -31,7 +31,7 @@ public class ReplyRestController {
 
     @DeleteMapping("/{replyId}")
     public ResponseEntity<Boolean> deleteReply(@PathVariable long replyId) {
-        boolean result = articleService.deleteReply(replyId);
+        boolean result = replyService.deleteReply(replyId);
 
         return ResponseEntity.ok(result);
     }
