+++ src/main/java/codesquad/springcafe/replies/model/Reply.java
@@ -1,6 +1,5 @@
-package codesquad.springcafe.articles.model;
+package codesquad.springcafe.replies.model;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
