+++ src/main/java/codesquad/springcafe/users/repository/H2UserRepository.java
@@ -4,7 +4,6 @@ import codesquad.springcafe.users.model.User;
 import codesquad.springcafe.users.model.data.UserCredentialData;
 import codesquad.springcafe.users.model.dto.UserPreviewDto;
 import codesquad.springcafe.users.model.dto.UserUpdateDto;
-import codesquad.springcafe.users.model.dto.UserUpdateRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
